dedicated debug_api.go for DebugAPI
